Move tests to subfolders
It was already getting a bit annoying to have to keep track of which test-files were integrationy and which just unittests. Now it’s clearer.
